edited task-3, changing method

+++ js/task-3.js
@@ -1,7 +1,7 @@
 'use strick'
 
 function getElementWidth(content, padding, border) {
-    return Number.parseInt(content) + Number.parseInt(padding) * 2 + Number.parseInt(border) * 2;
+    return Number.parseFloat(content) + Number.parseFloat(padding) * 2 + Number.parseFloat(border) * 2;
 }
 console.log('TASK - 3');
 console.log(getElementWidth("50px", "8px", "4px")); // 74
